refactor: use direct WhatsApp helpers
